requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.10
google-api-core==2.11.0  
google-auth==2.19.0  
google-cloud-bigquery==3.10.0  
google-cloud-storage==2.8.0 
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
db-dtypes==1.1.1  
fastavro==1.10.0